pyproject.toml
[tool.poetry]
name = "4Challenge"
version = "1.0.0"
authors = ["Enkonix Dev Team <dev@enkonix.com>"]
description = "4Challenge"

[tool.poetry.dependencies]
python = "^3.6"
toml = "^0.10"

django-debug-toolbar = "^1.11"
django-environ = "^0.4"
django-handy = {git = "https://github.com/tatarinov1997/django-handy"}
django-redis = "^4.10"
djangorestframework = "^3.10"
flake8 = "^3.7"
gunicorn = "^20.0"
psycopg2-binary = "^2.8"
redis = "^3.4"
Django = "^2.2"
Pillow = "^7.1"
whitenoise = "^5.0"
mrfh = "0.0.1"
celery = "4.2.*"
billiard = "3.5.0.4"
safety = "^1.8"
wemake-python-styleguide = "^0.14"
pyyaml = "^5.1"
drf_yasg = "^1.16"
django-cors-headers = "^3.2"
sentry-sdk = "^0.14"
django-filter = "^2.2"
dotenv-linter = "^0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
